Opening multiple REPLs works much more nicely.
Previously, opening multiple REPLs or opening and
closing the REPL multiple times resulted in lots
of empty buffers being created, this no longer
happens. Also the first line was overwritten with
a prompt. If the buffer already exists a prompt is
written to the last line and the cursor jumps to
the correct position.