Removed arbitrary precision numbers from TODO list.
Investigated using the decimal class but it
doesn't seem to fit in well with vimcalc. For now
I think using floats and converting to ints
occasionally is reasonable for a first release and
a lot of testing has been invested in this.

TODO: Look into using arbitrary precision in a future
release. Will need to remove the conversion to
ints.